backend/requirements.txt
crewai==0.28.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
httpx==0.25.1
pydantic==2.4.2
python-multipart==0.0.6
langchain==0.0.340
facebook-business==18.0.5
pillow==10.1.0
websockets==12.0
numpy==1.26.2
pandas==2.1.3
matplotlib==3.8.2 